COMP: Specify std:: prefix for pow
To address:

/FdWrapping\Modules\MorphologicalContourInterpolation\CMakeFiles\MorphologicalContourInterpolationPython.dir\ -c Wrapping\Modules\MorphologicalContourInterpolation\itkMorphologicalContourInterpolatorPython.cpp
c:\projects\itkmorphologicalcontourinterpolation\include\itkMorphologicalContourInterpolator.hxx(149) : error C2668: 'pow' : ambiguous call to overloaded function

+++ Modules/Filtering/MorphologicalContourInterpolation/include/itkMorphologicalContourInterpolator.hxx
@@ -144,9 +144,9 @@ MorphologicalContourInterpolator<TImage>::MorphologicalContourInterpolator()
   , m_UseDistanceTransform(true)
   , m_UseBallStructuringElement(false)
   , m_UseCustomSlicePositions(false)
-  , m_MinAlignIters(pow(2, TImage::ImageDimension))
+  , m_MinAlignIters(std::pow(2., static_cast<int>(TImage::ImageDimension)))
   , // smaller of this and pixel count of the search image
-  m_MaxAlignIters(pow(6, TImage::ImageDimension))
+  m_MaxAlignIters(std::pow(6., static_cast<int>(TImage::ImageDimension)))
   , // bigger of this and root of pixel count of the search image
   m_ThreadCount(MultiThreader::GetGlobalDefaultNumberOfThreads())
   , m_LabeledSlices(TImage::ImageDimension) // initialize with empty sets
